requirements.txt
bech32==1.2.0
fastapi==0.78.0
joblib==1.1.0
mongoengine==0.24.1
python-dotenv==0.20.0
python_dateutil==2.8.2
requests==2.27.1
starlette==0.19.1
uvicorn==0.17.6
websocket_client==1.3.2
dnspython==2.2.1

